backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Code Analysis
radon==6.0.1
lizard==1.17.10
pygments==2.18.0

# Pattern Matching & Parsing
sqlparse==0.5.1

# AI Integration
openai==1.54.0
anthropic==0.39.0
google-generativeai==0.8.3
httpx<0.28  # Required for OpenAI compatibility

# Utilities
pydantic==2.9.0
python-dotenv==1.0.1
aiofiles==24.1.0
